Use a wrapper to the less engine to set the import relative path
This is a little wrapper class to the less engine.  I use this to access
and set the `@path` instance variable on the engine.  This sets the root
path all imports are done from.  This just makes importing partials easier.

I also updated the less dependency to only allow version that have
this engine implementation.