get working in modern ruby versions
This updates the gem dependencies to bring in the latest versions
(which also now work in modern ruby versions) and updates the test
suite to get up to convention for working in modern ruby versions.
This includes backfilling Array#sample so it works in 1.8.7.

Note: I also reworked the support Factory to be more up to our
latest conventions.  This caused no behavior changes, though.  I
also updated this gem to our latest conventions overall (ie no
rake deploy tasks, etc).  This is possible now that we have ggem
1.8.